--- calibrate_and_validate.py
@@ -1,11 +1,0 @@
-#!/usr/bin/python
-import calibrate
-import validate
-import validate_for_plot
-import shutoff
-
-from time import sleep
-
-if __name__=='__main__':
-	validate.main()
-	validate_for_plot.main()
